feat: automatically match system theme
This commit adds a small piece of logic to check whether the user has
specified Dark mode in their system theme, and sets the default value of
the theme accordingly.

We're using `matchMedia` which should be supported by all major browsers
so there shouldn't be any compatibility problem, but we do a quick check
for it and fallback to Light mode just in case.

	https://caniuse.com/?search=matchmedia

For the CSS selector itself, it's only IE that doesn't seem to support
it but I think that's fine

	https://caniuse.com/?search=prefers-color-scheme