feat: author searching
This commit adds commit searching based on the name of the author,
utilising fuzzy matching.

I initially thought about just adding a basic search, but quickly
realise that GitHub usernames can be a pain to type, so any help getting
to the search destination faster is a bonus - hence the fuzzy matching.

I didn't do too much digging into the fuzzy searching library, but
Fuse.JS seemed to be a sensible choice. The trade-off for performance is
a complex balance of the setup, search and teardown times.