fix selected users updating on search
We were recreating the list of users that we pick for deployment on each
render and, due to the way the input works, each keypress causes a
render.

This means that we need to memoise `developerCommits` so that it's only
recreated when the relevant information changes (or on refresh).

However, we _also_ need to memoise `<DeploymentUsers />` otherwise the
order of the users _could_ change with each render.