feat: add show own commits button
I thought of a few ways about implementing this, but figured the
cleanest was to just add the login data that we already have as an
`endAdornment` on the `<TextField />`.

In the event that we don't have `loginData` (for whatever reason), we
just don't show the button - but this doesn't affect any of the existing
functionality.