feat: tweak the fuzzy search to be less fuzzy
It turns out that for the use-case of searching for names, the defaults
that come with Fuse.JS are a little bit _too_ loose. Therefore, this
tightens them up a bit by decreasing the threshold and distance as
outlined here:

    https://www.fusejs.io/api/options.html#fuzzy-matching-options

This reduces them down significantly, but I think that should be better
for match on exact names when the user clicks on the appropriate button.